feat: move the `join_set` tracing utils to a dedicated file
